Test alias in v1/rego
